LayoutTests/fast/profiler/one-execution-context.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("One Execution Context");
function testStart()
{
    endTest();
}

</script>
</head>

<body onload="testStart()">
This page has a single script tag with some simple JavaScript.
<br>
<br>
To use this test, load it in the browser then load the WebInspector and look at
the profile.  In the profile you should see one (program).
</body>
</html>
