LayoutTests/fast/profiler/one-execution-context-expected.txt

CONSOLE MESSAGE: line 59: One Execution Context
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) one-execution-context.html 8
CONSOLE MESSAGE: line 75:   (program) one-execution-context.html 4
CONSOLE MESSAGE: line 75:   onload one-execution-context.html 17
CONSOLE MESSAGE: line 75:    testStart one-execution-context.html 10
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page has a single script tag with some simple JavaScript. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile you should see one (program).
